string: Cast memchr and memrchr val to 'unsigned long' while building mask
In the speed-optimized version of this code, the needle is replicated
to fill an unsigned long. It does that by oring various shifts of the
value. Of course, C performs the first of those by first casting to
int, and on MSP430, the compiler ends up sign extending so values >=
0x80 end up with high bits set which wrecks the resulting mask value.

+++ newlib/libc/string/memchr.c
@@ -83,7 +83,7 @@ memchr (const void *src_void,
          character and then detecting for the presence of NUL in the
          result.  */
       asrc = (unsigned long *) src;
-      mask = d << 8 | d;
+      mask = (unsigned long) d << 8 | (unsigned long) d;
       mask = mask << 16 | mask;
       for (i = 32; i < sizeof(mask) * 8; i <<= 1)
         mask = (mask << i) | mask;
